Ensure logging of modules that failed to gather telemetry.
Calling `console.error` from within the `page.evaluate` callback logs
the error to puppeteer instances console not the console that is
actually running the codemod process.

This fixes that issue (so the messages actually bubble up now) by
using `page.exposeFunction` which **does** have access to the current
node process.